+++ packages/ui/src/components/Calendar/components/CalendarHeader/CalendarHeader.tsx
@@ -1,13 +1,8 @@
 import { FC, HTMLAttributes } from 'react';
-import style from './style.module.css'
-import { csx } from '@vega-ui/utils';
+import { CalendarBaseHeader } from '../../../CalendarBase';
 
 export type CalendarHeaderProps = HTMLAttributes<HTMLDivElement>
 
-export const CalendarHeader: FC<CalendarHeaderProps> = ({ children, className, ...props }) => {
-  return (
-    <div className={csx(style.header, className)} {...props}>
-      {children}
-    </div>
-  )
+export const CalendarHeader: FC<CalendarHeaderProps> = (props) => {
+  return <CalendarBaseHeader {...props} />
 }