+++ packages/utils/src/date/getWeekDayNames.ts
@@ -3,7 +3,7 @@ import { getFirstDayOfWeek } from './getFirstDayOfWeek';
 export const getWeekDayNames = (
   locale: string = 'en-US',
   style: 'long' | 'short' | 'narrow' = 'long',
-  firstDayOfWeek: number
+  firstDayOfWeek?: number
 ): string[] => {
   const baseSunday = new Date(Date.UTC(2025, 0, 5))
   const formatter = new Intl.DateTimeFormat(locale, { weekday: style, timeZone: 'UTC' })